:bug: Fix issue around departure and arrival location being the same.
The order of attributes passed to `locationIdOrNullIfOverlapping` mattered.
I was returning the wrong location id from the method.

Also add back-end validation to blow up if this happens since it's better to force the UI to fix itself than permit invalid data in the database.

+++ api/src/services/travel-segments/bulk-replace-service.ts
@@ -36,6 +36,18 @@ export class BulkReplaceService extends BaseService {
       throw new Error("All travel segments must have the same isActual value.")
     }
 
+    if (
+      this.travelSegmentsAttributes.some(
+        (travelSegmentAttributes) =>
+          travelSegmentAttributes.departureLocationId === travelSegmentAttributes.arrivalLocationId
+      )
+    ) {
+      const prettyTravelSegmentsAttributes = JSON.stringify(this.travelSegmentsAttributes)
+      throw new Error(
+        `Departure location and arrival location must be different for each travel segment: ${prettyTravelSegmentsAttributes}`
+      )
+    }
+
     return transaction(async () => {
       await TravelSegment.destroy({
         where: {
